requirements.txt
Flask==1.1.2
line-bot-sdk==1.13.0
pandas==0.25.1
gunicorn
